Add functionality to switch leaders
Moved leader attribute from the player to the game. When
clicking on a user name, the leader assignment will change.

TODO: President and Bomber assignment is not working correctly. The assignment
should be fired off in the assign roles function and assignment logic needs to be fixed.
One some loads, with only two players, there will be a bomber and one blue team member. That
member should be the President.

TODO2: Add roles with the blue cards in a show/hide toggle.

TODO3: Add rounds based on player numbers